example: Reorganize module
Reorganize the Employees example by splitting it into two modules,
Employees and EmployeesData, the latter containing all definitions.

This should make the example less verbose and hopefully easier to grasp.
